Renames UploadSessionInfo to UploadSession and UploadSessionInfoRequest to UploadSessionRequest
